Re #27966 Print averaged delta in SANS beam centre finder
Change the SANS Beam Centre Finder to prints the average residual change
between two quarters. This allows for direct comparisons, previously if
a user increased the radius limit it would appear to perform worse,
despite actually doing better. This was because the additional number of
points was not accounted for.