Re #28119 Append bad index to out of range exception
Updates the IndexType to have a new method to convert to a string, this
reduction in strict type safety could mean developers could abuse
SpectrumIndex -> String conversions. However, it enables us to print a
string representation of an index, such as if we need to say index x is
out of range.

Previously, the at() access on the IndexProperty map would throw with an
invalid key message. This is useless for anyone who is trying to Mask a
range of specra (such as 500), as you have to guess if the min / max or
a middle value was bad.