RE #27593 Fix file paths when loading by run number
The muon GUI generates a path that can be given to the LoadMuonNexus
algorithm to allow it to check the archive and any selected data
directories. That file path was what was shown in the GUI even if it was
not an accurate way of showing where the loaded file is.

Now, the file path is pulled from the algorithm after the load, so is
much more accurate than previously.