Make the help pages use the custom Mantid style sheet
The help pages in MantidPlot and MantidWorkbench were not using the custom Mantid
style sheet. This could be seen by the fact the hyperlinks were blue for example
instead of green.
These changes force the Qthelp pages to use the bootstrap theme by setting the
options qthelp_theme and qthelp_theme_options which seem be be new in recent
Sphinx versions. Previously Qthelp just shared the html_theme\html_theme_options
settings with the Html help.