Set flags on figure windows to keep them on top of mainwindow
We want plot windows to remain on top of their parent.
The old method of forcing the figure window to the top
when mainwindow gained focus ends up causing a fight with
the window manager over what has focus. Some window
managers, GNOME, allow the script editor to have focus, others
such as Xfce do not. This is a compromise as we lose the
minimize button (but not on Windows) yet the window is
more likely to play correctly with the rest of the ecosystem.

Different flags/parent behaviour is set per OS as each handles
this quite differently. The downside on some Linux window managers
is that the figures are now missing a minimize button. The visible
button in the Plots toolbox still works though.
Refs #27883