Re #28266 Fix progress arrow for Workbench
Fixes the progress arrow which previously only pointed at the first
line, irrespective of selection. Additionally removes a method which is
completely unused in the code-base, but looks like it would do a similar
thing as it's misleading.

A struct is used in case we need to add additional fields, and for the
extra readability when we dereference the map iterator. This avoids a
iter->value.at<0> in favour of having named fields of a struct.