Use cmake generate_mantid_export_header to generate DllConfig.h
This is being done in stages. I've made this change on the Algorithms project
The change consists of removing the manually created DllConfig.h header and replacing
it with one generated during the build by cmake. The MANTID_ALGORITHMS_DLL macro defined
in DllConfig.had was already partially used in the Algorithms project - so some classes
had the attribute DllExport and others already used MANTID_ALGORITHMS_DLL.
I've made it so all classes now use MANTID_ALGORITHMS_DLL.
The DllExport macro was defined in Kernel\System.h so I've removed this include from a
lot of files and replaced it with MantidAlgorithms\DllConfig.h.

As an aside the client projects that use Algorithms such as AlgorithmsTest were working
without this change so it appears that having DllImport present on a class definition
(which is what MANTID_ALGORITHMS_DLL expands to in client executables) is optional