Update docs and add attenuation profile file to 2 algorithms
While the primary use of the new attenuation profile file feature is expected
to be via the sample environment xml file, it also makes sense to add it into
any algorithms where the material properties can be specified. So I've added a
new attenuation profile parameter to the SetSampleMaterial and LoadSampleEnvironment
algorithms. The validations on these algorithms means they can't be called using
postional arguments so have inserted the new parameter in the logical place next
to the absorption cross section

Have updated the docs to explain how the attenuation profile file is used