Updated return value of remove_workspace_artists Re  #28434
This function returns a bool which was previously being used to indicate two separate things. The return values could mean:
  * False: Either the workspace does not exist on the axis or it does but the axis is not empty after its' removal.
  * True: Workspace exists and was removed and axis is now empty

This made the value not very useful for determining whether a call to remove_workspace_artists had removed anything. A grep of the code base returned that this return value was not currently being used anywhere except in MantidAxes.cpp where it was assumed to indicate whether anything had been removed. As you can easily check whether an axis is empty by other means I have changed the return value so that it indicates whether the call resulted in any change to the plot which is harder to divine by other means and brings this function in line with replace_workspace_artist.