Improve the stitch params tooltip on the reflectometry GUI
Previously this tooltip was misleading because it was referencing the transmission stitch params property of the reduction algorithm. This input is actually for all properties for the Stitch1DMany algorithm so we cannot take the documentation from a single algorithm property as we do for the other tooltips. Therefore, this commit adds a function to be able to set a tooltip independently of any algorithm and sets this tooltip manually.