Fix a bug where adding a new batch overwrites other batches
The problem here was that when the instrument is set on the new batch, a notification is sent that the instrument has changed and all batches were being reset. However, although the instrument has been set, we can just check whether it is different to the current instrument to check whether it really has changed, and if not, do not do the updates.

Re #28429