Fix a bug where restoring defaults affects all batches
Clicking restore-defaults on an Experiment or Instrument tab in a batch should only affect that tab and batch. This commit fixes a bug where a notification was sent to all batches, and every Experiment and Instrument tab was being reset. The fix is to separate updating the instrument (which is required to ensure we have up-to-date settings loaded from the instrument) from the subsequent notification which updates all batches. Instead, the particular Experiment/Instrument tab where the request was made restores its own defaults itself.

Re #28429