Switch m_quoteValue in IFunction to not rely on undef behaviour
m_quotevalue of the attributes did not explicitly set or copy the value
of m_quotevalue. This means that the memory was undefined but it worked
as we were doing a replacement in-place

However, this can randomly blow up on us if the heap is used in between
two calls, since the address of m_quoteValue will have changed.