Improve extrapolation beyond range of supplied attenuation profile
Previously, if an attenuation factor was requested for a lambda that
is outside the range of an explicit attenuation profile the code
extrapolated based on the gradient of the two most extreme points.
If the data was noisy this gave some poor results. Improve this by
falling back on the tabulated linear absorption coefficient of an
optional ExtrapolationMaterial