Add helper file for parsing an ISIS journal file
Adds a couple of generic functions for requesting data from a journal. The main work is done by a function that takes the file contents as a string so that this can be easily tested. A wrapper function allows the file to first be fetched from a URL constructed for the instrument and cycle.

Also adds a couple of functions for requesting the list of cycles, again with one function taking the file contents and another which finds the URL for the instrument.

Re #26897