Add external plotting button to the Muon Analysis GUI
This commit does the following things:

(1) Sets the parent of the Muon Analysis GUI to be the main workbench window. This prevents the external plot sending the interface to the back.

(2) Moves the number_of_axes variable to the view.

(3) Adds in an external plotting button within the Plotting Widget View.

(4) Creates an external plotting model view class. This class delegates to either workbench, or mantidplot based plotting.

(5) Adds unit tests for the new feature