Make KafkaDecoder tests run Lockstep to fix races
Converts the Kafka[Event/Histo]StreamDecoder tests to run in lockstep
whenever possible. The previous version held the unit test for n number
of Kafka steps then let both threads run free.

This fixes numerous races resulting in flakey tests, for example is
HistoStreamDecoder we waited for 1 iteration whereas three was required.
Rather than just tweaking the values and hoping this new mechanism will
deadlock if the threads get into a race-y condition. A timeout ensures
that after 2 minutes we fail with an abort stating we deadlocked.

This flushed out multiple races in the StreamDecoder too, however some
tests which rely on the EndOfRunStream have to run unsyncronised: The
the thread will block and the Kafka thread waits for ExtractData to be
called, causing a deadlock. If we are having problems in the future that
would be the first place to look, however for this commit it's out of
scope.