Abins VASP support: support XML extension, warn against OUTCAR
For validation/troubleshooting purposes it is useful to refer to the
human-readable OUTCAR file, which contains the frequencies that users
will expect to see reflected in the outputs of Abins.

Efforts have been made to have a robust OUTCAR parser that will raise
an error if the format changes. However, the .xml file output which is
simultaneously produced in VASP calcultions is a much better target as
the parsing is more likely to survive VASP updates.

In order to recommend the XML parser to users while keeping access to
the OUTCAR parser we:

- only refer to XML files in the main Abins doc page;

- explain in the concepts/AbinsImplementation docs (which are
  essentially Abins-specific developer notes) that the OUTCAR parser
  is for development purposes and may be deprecated; and

- raise a warning message to this effect whenever the OUTCAR parser is used.