Remove race between Test and Kafka on extractData
Resolves the race between the test thread and Kafka thread when Kafka is
waiting for extractData to be called. We now use the callback mechanism
in waitForRunObservation.

Also adds some mutex's so the kernel can see when a CV should be
rechecked, atomics work because eventually the scheduler gives in and
tries again (hence we need pred.s). However, if we protect the atomic
access the other thread is resumed when the lock is released from the
original thread.