Abins: VASP modifications for IBRION=6
A new VASP benchmark calculation was performed with the IBRION=6
calculation option. This uses finite displacements instead of
perturbation theory, permitting the use of a dispersion-correced
PBE-D3 functional.

This revealed an assumption in the vasprun parser that the dynamical
matrix ("dynmat") data would be included in the first (and only)
"calculator" element of the XML tree. In fact, the IBRION=6
calculation produced multiple such elements, and the "dynmat" block is
located in the last one. To account for this, the XML parsing has been
refactored to search through all calculator blocks, using the last
"dynmat" encountered.

Why not break the loop after encountering "dynmat"? Well, VASP allows
one to use these Hessian-calculating modes as geometry
optimisers (setting NSW > 1 in combination with IBRION = 5-8). A
sufficiently enthusiastic DFT practitioner with access to generous CPU
walltimes might thus have performed multiple vibration calculations in
the course of a single VASP run. This is rare as Quasi-Newton
pre-optimisation is generally recommended, but it would be a
nasty thing to troubleshoot later.