Abins VASP Loader: refactor generator for Python 3.7
PEP 479 is introduced by Python 3.7, and discourages the use of
StopIteration by having them raise errors rather than quietly
terminate a generator.

This was really just being used as a kind of ``break`` to avoid some
indentation, so just swap the logic around to remove the explicit
termination. The other option is to use ``return``.