Fix missing reference to the instrument scroll bar
This commit adresses the integration scroll bar on the instrument render
tab. A previous commit deleted this bar altogether when it was not
needed, mainly for monochromatic workspaces. But a few methods,
especially in MantidPlot, actually depends upon the attribute being
initialized.

As such, the integration scroll bar object is again initialized in every
case, but a flag "hidden" have been added, which allows it to be created
as a mere shell, without any component that would show onscreen.