Reverted changes made to how data was added to models Re #28057
These changes were originally made to simplify the interface of how data was added to the model and facilitate easier testing. This led to some overrides being ignored however breaking the ability to add workspaces in multi input mode for the FQ and convolution tabs. As this PR is just supposed to be removing uneeded code. I have reverted these changes for now and removed the tests they enabled.