WIP Abins: overhaul file/module names for Python library
The existing naming and usage conventions for Abins are quite
confusing. Classes are defined in module files sharing a TitleCase
name, and imported by __init__ into the AbinsModules namespace. This
makes it very difficult to reason about when you are looking at a
class and when you are looking at a module. In an especially
misleading example, AbinsModules.IOModule is usually not a module but
a *class*.  The primary benefit of this scheme is that it is easy to
determine in which _file_ classes are defined.

The PEP8 recommendation is that python module names should be
lowercase and short, with underscores for clarity if necessary. Here I
adopt a consistent pattern that underscores are only used if the
module is _not_ named after a class (e.g. test_helpers); modules that
define classes are named with a lower-case rendition of the TitleCase
class name. This mostly preserves the ease of finding appropriate
files, while clearly distinguishing modules from classes.

There is a lot of redundant use of the name "Abins" for classes that
are already within the Abins namespace. This has been streamlined,
such that e.g. `AbinsModules.AbinsParameters` becomes
`abins.parameters`. As a result, this commit removes over 150 instances of
the word "abins" from the Mantid codebase.