Abins: remove single-crystal modules and tests
Abins includes placeholder code for unimplemented single-crystal
functionality. The proposed implementation is not documented.

Work on simulation of single-crystal samples is on the roadmap for
Abins, but is likely to involve integration with the external Euphonic
library, developed in collaboration with the Excitations group at
ISIS. It is not clear how useful these draft implementations will be
as part of such an implementation. If they are helpful, they can be
rescued from the Git history. Before then, there is little point in
maintaining this unused code for distributed Mantid releases.