Abins: Rename "CalculateX" classes to "XCalculator(Factory)"
The CalculatePowder and CalculateS components of Abins are
misleadingly-named.

CalculatePowder is a _class_, not a function or method, that
_performs_ (or at least calls) calculations. The name
should be a noun: PowderCalculator.

CalculateS is a class that primarily provides a method to dispatch an
appropriate constructor and return a calculator. Such classes are
generally referred to as "factories", so we rename it
SCalculatorFactory. The increased verbosity is unfortunate, but it
makes it clearer that lines in Abins that "CalculateS.init()" are not
actually calculating anything, but constructing a calculator.