[DRILL] change color of rows in the table widget
New functions to change the background color of rows have been added to
the drill table widget. This allows to display the currently processing
row, the processed rows and the failed rows.
When signaling the end of a process, the model sends also the reference
of this row. Same happens for the failed processes.