Stop regenerateTracksForEachLambda being hardcoded
Correct mistake from commit d48169a8349060223910e0ff7524ba95c17bfdaa
where the regenerateTracksForEachLambda parameter of MCAbsorptionStrategy
was left hardcoded to False following the creation of a user visible
parameter on the MonteCarloAbsorption algorithm which allowed the user to choose.

Setting the user visible value to True was having one minor impact on the calculation
- for calculations using a sparse instrument it was causing a sparse wavelength
grid to be used during the simulation.

Have adjusted some unit tests for tests that were using a value of "true" while
actually running the calculation with false. This was either a change to the
values being checked or in some cases I've simply switched the test to use the false
value because this is the recommended value