Initial fix for invalid nGEM data file load
The data file concerned had its first event chopped, this
patch solves this issue by seeking forwards in the file until a
valid event is located. I would be interested to know if this
chop is a bug or feature i.e. is the other half of the event left in
another data file and what should be done is concaternating the
buffers for a multi-file load.