Rename Abins loaders for external data
Ab initio file loaders are classes/objects and their names should be
nouns. The current notation is misleading in that e.g. instantiating
"LoadCASTEP" does not, in fact, load anything.

The loaders are moved into a new submodule abins.input. This is
slightly dubious as `input` is a python built-in, but no serious
problems are anticipated as API users would generally

  import abins.input`

or, more likely,

  from abins import CASTEPLoader

(as the loader objects are also accessed in abins.__init__).

Still, alternative name suggestions are welcome.