RE #28052 Fix overplotting in colorfill plots
Colorfill plots would fail to open the figure options if the plot had
been overwritten by dragging and dropping another workspace over it to
overplot it. No longer passing the axes object to the plotting of the
colorfill, forcing a new one to be created on the figure seems to
have fixed the issue.