Back out one of performance improvements and release GIL
The performance improvement in LoadNexusLogs.cpp was causing a problem with
the processing of the period_log integer time series log entry. So I've removed this

I have also added a line to the getPropertyAsSingleValueWithDefaultStatistic function
in Run.cpp to release the GIL. This stops the call from Python blocking while the
sample log averages are calculated. Reduces appearance of a hang from the end user
perspective