Hide the wb logs dialog plot and stats if there are no suitable logs
So if there are no time series logs that are plottable, then the stats and plot section of the window will be hidden.
This should be dynamic so that for MD workspaces with multiple experiment infos
It will disappear and reappear as the experiment info is changed.

Although I can't think or a situation where you would have a md workspace like that.