Use a range for the bin plot x-axis when possible
When the monitors are at one (or both) end of the spectra's list, it is
possible to use a range rather than a list to describe the indices of
the detectors. For big workspaces, it is significantly faster.
However, when it is not possible, i.e. when the monitors are in the
middle of the spectra list (which is a case that should not happen
often), the detectors' indices are stored in a numpy array.