Fix a bug loading old -add files
"Add" files (which have a -add.nxs suffix) have a 'workspace_name' entry in the nexus file when they are generated by the new ISIS SANS GUI. There is a check on the workspace name format when checking the file type (specifically if it is added event data). However, -add files produced with the old GUI do not have the 'workspace_name' entry and the new GUI throws an error when it does this check, even for histogram mode where it is not relevant. This fix checks if the entry exists first and if not uses the file name instead.

Re #28652