Apply changes from code review
This commit changes a comment in pdffouriertransform2.
Takes direction from the input parameter instead of ws unit.
Removes references to Q and R in terms of input and output.
Sets filter to false on pdffouriertransform for the fourier filter in polaris.
Moves the filter and rho0 property from the groups to being independent.
Fixed an issue with sample details being lost on transformation.

re: #28596