Convert PSIBackgroundSubtraction to a data processing algorithm
This commit converts the PSIBackgroundSubtraction algorithm to a data processing algorithm. This increases the clarity and transparency of the workspace history.

In order to do this, the operator overload which removed the background has been replaced with a CreateWorkspace call, which generates a workspace based on the backgrounds obtained, followed by a call to the child algorithm "Minus".