Added centre as option in delta function.
A delta function is defined by its centre and height. Previously only the height value was displayed in the simplified template function browser as it was assumed the centre would almost always be at 0 or close to it. Having taken scientific feedback however it was decided to add the centre as an option as well.