[DRILL] sans settings dialog
The SANS settings dialog has been modified. It contains only one tab
with widgets to collect all algorithm parameters.
When processing, the settings are used in association with the row
parameters to process the data. When custom options are provided (in the
table), they are used to update the global settings for this particular
row.