Fix home button when line plots are activated
The navstack in the toolbar stores boundingboxes
of the axes but as we have resized the axes these
positions are no longer correct to get the image
data back to displaying all of the data. We
have no back/foraward buttons so the stack
is not that useful. Implements a simple reset to
data limits when home is clicked.