Set maximum number of threads
After profiling the loader it was observed that CPU usage jumped to 100% in the LoadISISNexus loader. This is unnecessary usage given the small size of the Muon Nexus files. This lead to slow performance, where performance was between 10-100% slower than LoadMuonNexus1. With this change the speed difference is around 10% and with much less variability.