Abins: Fixes and cleanup following rebase
Abins recently went through a large refactor and the VASP Loader was
rebased onto these changes. A few issues arising from the automatic
rebase are corrected here, along with some accompanying tweaks to the
import structure.

In general there is a lot of interdependence between the Abins, giving
a high risk of import loops. The likelihood of these has been slightly
reduced by removing abins.input from the automatically-imported
submodules in abins.__init__. The main remaining issue is that
AbinsData, a high-level generic data container, contains references to
all the file loading classes so that it can dispatch to them. This was
a cause of import loops, and the immediate issue has been mitigated by
deferring these imports until the relevant method is called. However,
we should note that in general this is not going to scale well and we
should not be importing more modules as more parsers are added. At
present, if importing any Loader fails then _all_ the Loader tests
will fail (as they depend on AbinsData). The solution may be to use
importlib to programmatically grab the appropriate module as needed,
rather than storing all of them in a dict.