Fix error when changing normalisation with log colorbar
A ragged workspace with a log colorbar could end up with vmin and vmax
values being negative, which is incompatible with a log scale. The
change_colorbar_axes method already has handling for this, so the
normalisation toggle sets the colorbar to normalised before adjusting
the data and then puts it back afterwards using the handling from the
existing method.

RE #28391