Improving handling of workspace deletion
This commit addresses the sequential fitting tab not behaving correctly when a workspace is deleted.
This was fixed by connecting the tab to the plot deletion signal in the context. This brought up another issue with the creation of empty tables in the sequential widget, which has also been fixed.

A unit test was also added.