Avoid creating peaks presenter when not required
The lazy instantiaion of the peaks presenter meant
that the first call to the property created the
object even in instances where the peaks button
had not been pressed or it was disabled!
This manifested itself on the view for matrix workspaces
constantly responding to resize events that inserting
an empty peaks view would create.
Refs #28579