Fix LoadMuonNexusV2 wrongly loading PSI Muon files
The LoadMuonNexusV2 algorithm was wrongly being chosen by the Load algorithm to load PSI Muon files. This occurs as the files are very similar, with both having a /raw_data_entry and a raw_data_1/definition Muon entry.

In order to differentiate these two files, the MuonNexusV2 loader now has an additional confidence metric, which returns 0 if a top level beamline entry is not found, i.e raw_data_1/beamline. This works as the beamline entry in the PSI files is located within raw_data_1/instrument/beamline entry.