Reset view limits on new matrix plots
If the dimensions are flipped while the view is zoomed
or panned then it is unclear what area a user would
wish to see and its possible the new area is outside
of the image. For this reason reset the view limits
to the data limits on making a new plot.