Re #28255 Change initial values of a and b for IkedaCarpenter
This commit changes the initial values set for alpha0, alpha1 and beta0 in the ikedacarpenter function to depend on the x units so the initial plot guess is reasonable.
It also adds a check so that when a peak function is added the width of the function is within the fitting region.