[DRILL] valition of each cell during table filling
Each time a value is added in the table, it is checked by the model
against the selected algorithm. If a value in not correct, the view
changes the color of the corresponding cell and add a tooltip message
to inform on the error.

Invalid cell values prevent the processing of the corresponding rows.
The User has to correct them before. When corrected, the color of the
corresponding cells is removed as well as the tooltip.