Create plot help pages and expose help window to python
This commit creates the help pages for workbench plots and exposes the help window to python.
Using this help window, a "Question mark" on the workbench plots can be pressed to open a relevant doc page.
To determine which plot type its dealing with,  the figure_type method is used.