[DRILL] change technique to acquisition mode
An instrument is linked to a specific technique and to a certain number
of acquisition mode(s). The selection of an acquisition mode is
responsible of the selection of a specific algorithm and parameters.

New methods have been added to the model to deal with instrument and
acquisition mode selection. The associated tests have been implemented.