Quality of life improvements for Function Dialog
This commit alters the following behaviour of the function dialog:

(1) The function box is initially clear so the user can type immediately 
(2) The peak type text is now a placeholder, representing the currently selected peak type.
(3) Complete of the function can be activated if enter is pressed and the list of possible completions is 1.