[DRILL] cut, copy, paste for cells
The cut, copy and paste actions are now valid for cells.

When cuting or 'copying, the selection has to be valid in order to
take the cells value in the local buffer (i.e. it has to have a
rectangular shape). It also must not concern only empty cells (this
should avoid erasing a whole set of cells when pasting).

When pasting, the selection shape has to match the buffer shape.
There is an exception if the buffer contains a single value; in that
case, this value will be repeated in every cells of the selection, no
matter its shape.