Prevent a shape being stored against sample pos
The shape of the sample should be stored against the m_sample field in the
ExperimentInfo class. This change prevents a second (potentially different)
shape being stored in the instrument definition on the "SamplePos" component

I've achieved this by requiring the "SamplePos" component to be of type
Component instead of ObjComponent. There were various places where the sample
was set up to be an ObjComponent without it needing to be so I've changed these
to Component

Have also adjusted bounding box behaviour. For a component without a shape
still make it contribute to the overall bounding box of the component list
based on its position. Also distinguish component at the origin without a
shape from a null bounding box