Add saving and loading function
When saving projects, the save has to register if the instrument window
has an integration scroll bar or not, so when it loaded the loader won't
attempt to add values to a non-existent bar, resulting in a segfault.
Also, it remembers the mask tube and mask pixel buttons being checked
and restores them to their previous state at loading, as for any other button.