Modify the integration scroll bar removal process
When not setting an integration scroll bar, instead of just not showing
it, rather not create it at all. This is mmore heavy handed, as it
implies checking for the existence of the object each time it is called,
but it is more reliable (as the non-created object cannot possibly
occupy any space on screen).

This solves the problem of the unclickable Render and Pick tabs.