Improve workbench handling of single valued workspaces
This commit adds a set of axis to the WorkspaceSingleValue such the data can now be displayed. It also updates the tree widget to account for this workspace type.

Lastly, it alters the behaviour of the workspacewidget to show data when double clicking on a single bin single spectrum workspace.